Fix: Separete ContentView in MainApp and added SplashScene
